refactor : Profanities, PurgomalumClient를 common 패키지 하위로 이동
