+++ src/main/java/kitchenpos/common/tobe/domain/Profanities.java
@@ -1,4 +1,4 @@
-package kitchenpos.products.domain;
+package kitchenpos.common.tobe.domain;
 
 public interface Profanities {
     boolean containsProfanity(String text);
