+++ src/main/java/kitchenpos/common/tobe/exception/WrongDisplayedNameException.java
@@ -1,4 +1,4 @@
-package kitchenpos.products.tobe.exception;
+package kitchenpos.common.tobe.exception;
 
 public class WrongDisplayedNameException extends IllegalArgumentException {
     public static final String DISPLAYED_NAME_SHOULD_NOT_BE_EMPTY = "DisplayedName은 비워 둘 수 없습니다.";
